Implemented new methods in Logic
Also put null checks in Create and Delete methods

+++ DDB2DA_HFT_2021221.Logic/DriverLogic.cs
@@ -17,13 +17,39 @@ namespace DDB2DA_HFT_2021221.Logic
             this.repo = repo;
         }
 
+        public void AddPoints(int id, double points)
+        {
+            repo.AddPoints(id, points);
+        }
+
+        public void ChangeId(int id, int newId)
+        {
+            repo.ChangeId(id, newId);
+        }
+
+        public void ChangeTeam(int id, int newTeamId)
+        {
+            repo.ChangeTeam(id, newTeamId);
+        }
+
         public void Create(Driver driver)
         {
+            if (ReadOne(driver.Id) != null)
+            {
+                throw new InvalidOperationException
+                    ("Creating new item with Id: (" + driver.Id + ") is not possible: this item already exists.");
+            }
+
             repo.Create(driver);
         }
 
         public void Delete(int driverId)
         {
+            if (ReadOne(driverId) == null)
+            {
+                throw new NullReferenceException("Driver with the Id: (" + driverId + ") cannot be found.");
+            }
+
             repo.Delete(driverId);
         }
